refactor: 관심사별 상수 분리
