+++ src/consts/common.ts
@@ -1,25 +1,5 @@
-export const DELAY = {
-  TOAST_HIDE: 1000,
-  TOAST_REMOVE: 2000,
-};
-
 export const URL = {
   MODE: 'mode',
   QUERY: 'query',
   PAGES: 'pages',
-};
-
-export const PAGE_SIZE = 20;
-
-export const LOCAL_STORAGE_KEY = 'userRating';
-
-export const RATING_STAR_SIZE = 5;
-
-export const RATING_SCORE = {
-  0: '평가해주세요',
-  2: '최악이에요',
-  4: '별로예요',
-  6: '보통이에요',
-  8: '재미있어요',
-  10: '명작이에요',
 };
