+++ src/consts/movie.ts
@@ -1,0 +1,5 @@
+export const MOVIE_LIST = {
+  TOAST_HIDE_DELAY: 1000,
+  TOAST_REMOVE_DELAY: 2000,
+  PAGE_SIZE: 20,
+};
