+++ src/consts/userRating.ts
@@ -1,0 +1,12 @@
+export const LOCAL_STORAGE_KEY = 'userRating';
+
+export const RATING_STAR_SIZE = 5;
+
+export const RATING_SCORE = {
+  0: '평가해주세요',
+  2: '최악이에요',
+  4: '별로예요',
+  6: '보통이에요',
+  8: '재미있어요',
+  10: '명작이에요',
+};
